use long as id in reposiotries

+++ src/main/java/my/photomanager/photo/cameraSettings/CameraSettingsRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import lombok.NonNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CameraSettingsRepository extends JpaRepository<CameraSettings, Integer> {
+public interface CameraSettingsRepository extends JpaRepository<CameraSettings, Long> {
 
 	Optional<CameraSettings> findByCameraModelName(@NonNull String cameraModelName);
 }
